Switch to using docker compose
Switched to using docker compose. The upshot is everything
boots up in one shot and changes to the python code
are reflected immediately in the flask debug server!

The bad news is I may have broken developing each of
these indepedently, but this is where I was able
to get to in the time I had allocated tonight.

I also found a bug where RP_BASEPATH doesn't
seem to be being included in the url anymore.